Refactor FeedbackWidget component and move it to a separate directory
